fix(update): retain activation ownership through finalization
Use the canonical Doctor policy for fresh finalization and package verification so only the updater activates the managed Gateway. Preserve named profiles and explicit stop/update/start guidance in recovery commands.

Move the web update lifecycle into its own owner and fence delayed hold responses from newer campaign or authority state. Cover the live-found activation defect and UI races with existing owner-boundary tests.
